Add new scripts for brain mask training and GAN model application; include image registration and subtraction functionality
Refactor UI layout and add new functionality for brain mask creation

This commit renames and moves the GAN and brain mask template files to a new "Future_scripts" directory. It also modifies the UI layout for the SEEG_masking module, changing the dimensions of a rectangle and removing some spacer items. Additionally, it adds new functionality for creating a brain mask, including input selection, applying a mask, and saving the mask. These changes improve the user experience and provide a foundation for future brain mask creation.